requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
m3o-py==0.1.1a1
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
python-dotenv==1.0.0
typing_extensions==4.5.0
yarl==1.8.2
